# win-x64/openapi.json
{
    "openapi":"3.0",
    "info":{
       "title":"My API",
       "description":"This is my API",
       "version":"3.0.1"
    },
    "servers": [
      {
         "url": "https://127.0.0.1:8000"
      }
   ],
    "paths":{
       "/":{
          "get":{
             "summary":"Test",
             "operationId":"test__get",
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            
                         }
                      }
                   }
                }
             }
          }
       },
       "/api/v1/tasks":{
          "get":{
             "summary":"Get Tasks",
             "operationId":"get_tasks_api_v1_tasks_get",
             "parameters":[
                {
                   "name":"title",
                   "in":"query",
                   "required":true,
                   "schema":{
                      "type":"string",
                      "title":"Title"
                   }
                },
                {
                   "name":"status",
                   "in":"query",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "title":"Status"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            "type":"array",
                            "items":{
                               "$ref":"#/components/schemas/Task"
                            },
                            "title":"Response Get Tasks Api V1 Tasks Get"
                         }
                      }
                   }
                },
                "422":{
                   "description":"Validation Error",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/HTTPValidationError"
                         }
                      }
                   }
                }
             }
          }
       },
       "/api/v1/tasks/{task_id}":{
          "get":{
             "summary":"Get Task",
             "operationId":"get_task_api_v1_tasks__task_id__get",
             "parameters":[
                {
                   "name":"task_id",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "title":"Task Id"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/Task"
                         }
                      }
                   }
                },
                "422":{
                   "description":"Validation Error",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/HTTPValidationError"
                         }
                      }
                   }
                }
             }
          },
          "put":{
             "summary":"Update Task",
             "operationId":"update_task_api_v1_tasks__task_id__put",
             "parameters":[
                {
                   "name":"task_id",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "title":"Task Id"
                   }
                },
                {
                   "name":"title",
                   "in":"query",
                   "required":true,
                   "schema":{
                      "type":"string",
                      "title":"Title"
                   }
                },
                {
                   "name":"status",
                   "in":"query",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "title":"Status"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/Task"
                         }
                      }
                   }
                },
                "422":{
                   "description":"Validation Error",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/HTTPValidationError"
                         }
                      }
                   }
                }
             }
          },
          "delete":{
             "summary":"Delete Task",
             "operationId":"delete_task_api_v1_tasks__task_id__delete",
             "parameters":[
                {
                   "name":"task_id",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "title":"Task Id"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            
                         }
                      }
                   }
                },
                "422":{
                   "description":"Validation Error",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/HTTPValidationError"
                         }
                      }
                   }
                }
             }
          }
       },
       "/api/v1/tasks/":{
          "post":{
             "summary":"Create Task",
             "operationId":"create_task_api_v1_tasks__post",
             "requestBody":{
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/Task"
                      }
                   }
                },
                "required":true
             },
             "responses":{
                "200":{
                   "description":"Successful Response",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/Task"
                         }
                      }
                   }
                },
                "422":{
                   "description":"Validation Error",
                   "content":{
                      "application/json":{
                         "schema":{
                            "$ref":"#/components/schemas/HTTPValidationError"
                         }
                      }
                   }
                }
             }
          }
       }
    },
    "components":{
       "schemas":{
          "HTTPValidationError":{
             "properties":{
                "detail":{
                   "items":{
                      "$ref":"#/components/schemas/ValidationError"
                   },
                   "type":"array",
                   "title":"Detail"
                }
             },
             "type":"object",
             "title":"HTTPValidationError"
          },
          "Task":{
             "properties":{
                "id":{
                   "anyOf":[
                      {
                         "type":"integer"
                      },
                      {
                         "type":"null"
                      }
                   ],
                   "title":"Id"
                },
                "title":{
                   "type":"string",
                   "title":"Title"
                },
                "status":{
                   "type":"integer",
                   "title":"Status"
                },
                "deadline":{
                   "type":"string",
                   "format":"date-time",
                   "title":"Deadline"
                }
             },
             "type":"object",
             "required":[
                "title",
                "status",
                "deadline"
             ],
             "title":"Task"
          },
          "ValidationError":{
             "properties":{
                "loc":{
                   "items":{
                      "anyOf":[
                         {
                            "type":"string"
                         },
                         {
                            "type":"integer"
                         }
                      ]
                   },
                   "type":"array",
                   "title":"Location"
                },
                "msg":{
                   "type":"string",
                   "title":"Message"
                },
                "type":{
                   "type":"string",
                   "title":"Error Type"
                }
             },
             "type":"object",
             "required":[
                "loc",
                "msg",
                "type"
             ],
             "title":"ValidationError"
          }
       }
    }
 }